# python-sdk/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@katanemo.com
    name: Katanemo Labs Inc.
    url: https://api.katanemo.com
  description: "With Katanemo developers can add support for users, enterprise SSO,\
    \ machine keys and fine-grained authorization in minutes."
  termsOfService: https://www.katanemo.com/terms-of-service
  title: "Katanemo - identity, and fine-grained authorization for modern { API-first\
    \ } software companies."
  version: 1.0.0
servers:
- url: https://api.katanemo.com
...
paths:
...
  /service/{serviceId}/tags:
    get:
      description: Gets tags associated with a resource of a service
      operationId: GetTagsForResource
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetTagsRequest'
        description: Tags and resource id
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'
          description: tags
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
      summary: Gets tags for a resource
      tags:
      - service
      - access-control
    post:
      description: "Add tags (key/value pair) to a particular resource that is created\
        \ for a service, for a particular organization account id."
      operationId: createTags
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tags'
        description: Tags and resource id
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'
          description: User account
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Add tags to a resource
      tags:
      - service
      - access-control
...
  /assignrole:
    post:
      description: Assign role to an identity principal
      operationId: AssignRoleToPrincipal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleObj'
        description: Role assignment
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: User account
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Assign role
      tags:
      - access-control
  /assumeRole:
    post:
      description: Creates a token with requested roleId
      operationId: AssumeRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssumeRoleObj'
        description: Role assignment
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: token
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Assume role
      tags:
      - access-control
...
  /org/{accountId}/role:
    get:
      description: Returns a list of all roles belonging to provided organization
        ID
      operationId: GetRolesForAccount
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoleResponse'
                type: array
          description: roles belonging to provided account ID
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List Roles
      tags:
      - organization
      - access-control
    post:
      description: Creates a new Role
      operationId: createRole
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
        description: Role to add to the system
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
          description: Role
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Creates Role
      tags:
      - organization
      - access-control
  /org/{accountId}/role/{roleId}:
    get:
      description: Gets a particular role for an organization
      operationId: GetRole
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: roleId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
          description: role
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get Role
      tags:
      - organization
      - access-control
    put:
      description: Update role
      operationId: updateRole
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: roleId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
        description: Role object that is being updated.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
          description: Updated service object.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update Role
      tags:
      - organization
      - access-control
...
  /token:
    post:
      description: Returns a short-lived token for API key/secret pair. Tokens contain
        claims that identify what a principal can or cannot do.
      operationId: getShortTermToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Generate Token for an authentication principal
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get Token
      tags:
      - access-control
  /oauth/token:
    post:
      description: Get an OAuth2.0 Token for an Authorization Code
      operationId: getOAuthToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthTokenRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenResponse'
          description: Getting token for client ID successful.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: OAuth Token
      tags:
      - access-control
...
components:
  schemas:
...
    Tags:
      example:
        accountId: accountId
        serviceIdPath: serviceIdPath
        resourceId: resourceId
        name: name
        serviceId: serviceId
        token: token
        tags:
          key:
          - tags
          - tags
      properties:
        serviceIdPath:
          title: serviceIdPath
          type: string
        serviceId:
          title: serviceId
          type: string
        name:
          title: name
          type: string
        resourceId:
          title: resourceId
          type: string
        accountId:
          title: accountId
          type: string
        token:
          title: token
          type: string
        tags:
          additionalProperties:
            items:
              type: string
            type: array
          title: tags
          type: object
      required:
      - name
      - resourceId
      - serviceId
      - tags
      - token
      title: Tags
    GetTagsRequest:
      example:
        accountId: accountId
        resourceId: resourceId
        resourceName: resourceName
      properties:
        resourceName:
          title: resourceName
          type: string
        resourceId:
          title: resourceId
          type: string
        accountId:
          title: accountId
          type: string
      required:
      - accountId
      - resourceId
      - resourceName
      title: GetTagsRequest
...